requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2 
django-cors-headers==4.3.1
python-decouple==3.8
dj-database-url==2.1.0
psycopg2-binary==2.9.9

# Google Generative AI
google-generativeai==0.3.0

# Firebase Admin SDK
firebase-admin==6.2.0

# Deployment
gunicorn==21.2.0
whitenoise==6.6.0

# Development
django-extensions==3.2.3
